+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.2] - 2019-12-18
+### Fixed 
+- Campaign expiration time support
+
 ## [1.1.1] - 2019-12-03
 ### Changed 
 - Case time support
@@ -25,7 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rewriting from Python to PHP
 
 
-[Unreleased]: https://github.com/adshares/adpay/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/adshares/adpay/compare/v1.1.2...HEAD
+[1.1.2]: https://github.com/adshares/adpay/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/adshares/adpay/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/adshares/adpay/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/adshares/adpay/compare/v0.1.0...v1.0.0